bup save could fail trying to read the metadata of a deleted file
If bup index see that a file foo has no change, bup save still need to read
foo in order to update the metadata. If the user removed foo between the
index and the save, the metadata call metadata.from_path would result in a
non intercepted error. This commit adds a try/except to handle this case.

Signed-off-by: Damien Robert <Damien.Olivier.Robert+...@gmail.com>
[git@tim-riemenschneider.de: rebased to current tmp/pending/meta:
since metadata is (now) stored in the index, only 1 of 2 hunks still applies
(race condition between reading the file for its content and stat-ing it for
metadata)]