Correcting a metadata bug when directories are removed
The preceding commit handled the case when a file was removed but still
present in an index. In this case, bup save tries to read the metadata of
the file, which gives an exception since the file does not exist anymore,
and the exception was not catched.

This commit do the same but for directories rather than files. The code for
directories is not at the same place than for the files, so the preceding
catch did not catch the exception corresponding to a removed directory.

I also added the catch in
  if first_root == None
although admittedly it would only produce if the whole structure that bup
save tries to save was gone, so it would not make much sense for the user
to save it...

Signed-off-by: Damien Robert <Damien.Olivier.Robert+...@gmail.com>