bup save creates a unreachable blob
When bup save has nothing to save (clean index), an unreachable
(unreferenced) blob is created. By the looks of it, it's a .bupm-file.

I would suppose, that bup save starts collecting metadata as it walks
the index and creates a .bupm-file. When there is nothing to save, no new
commit is created, but the metadata-blob is commited to the repository.

This commit merely contains a testcase (purposely put after all other
tests, because it will fail) to reproduce the problem.

Signed-off-by: Tim Riemenschneider <git@tim-riemenschneider.de>