Correct cache_get
Before, cache_get relied on the fact that every sub-path is in the cache,
ie. for each item in the cache, all parents are also in the cache.

If that is not the case, the cache-key is not calculated right, ie there
are sections missing. This patch corrects that.

Fuse should now be correct again (but slow)