Add Bloom filter based "bup gc" (UNSAFE)
WARNING: This code is UNSAFE, please test it with care.  Among other
things, it adds the first bup command that can modify packfiles.  Also,
the code assumes/requires a "quiet" repository; any simultaneous access
is dangerous.

Add a "bup gc" command that uses a Bloom filter to track the live
objects during the mark phase.  This means that the collection is
probabilistic -- that it may retain some (known) percentage of
garbage, but it can also work within a reasonable, fixed RAM budget
for any particular percentage and repository size.

In many cases, this may be sufficient and preferable, but in the
longer run, we might also want to add a precise method that guarantees
bup retains *no* unreachable/dead objects -- nothing here should
prevent that.

The collection proceeds as follows:

  - Scan all live objects by walking all of the trees rooted at any
    ref in git.list_refs(), and insert every hash encountered into a
    new Bloom "liveness" filter.  Compute the size of the liveness
    filter based on the total number of objects in the repository.  This
    is the "mark phase".

  - Clear the data that's dependent on the repository's object set,
    i.e. the reflog, the normal Bloom filter, and the midxes.

  - Traverse all of the pack files, consulting the liveness filter to
    decide which objects to keep.

    For each pack file, rewrite it iff it probably contains more than
    (currently) 10% garbage (computed by an initial traversal of the
    packfile in consultation with the liveness filter).  To rewrite,
    traverse the packfile (again) and write each hash that tests
    positive against the liveness filter to a packwriter.

    During the traversal of all of the packfiles, delete redundant,
    old packfiles only after the packwriter has finished the pack that
    contains all of their live objects.

The current code unconditionally tracks the set of tree hashes seen
during the mark phase, and skips any that have already been visited.
This should decrease the IO load at the cost of increased RAM use.
Thanks to Tim Riemenschneider for the suggestion, and an earlier
implementation.  Thanks to Ben Kelley for testing and reporting bugs in
earlier versions of this implementation.

Also note that until we add a "bup rm", the only way it should be
possible to cause the garbage collection to actually rewrite or delete
any packfiles is to manually manipulate the git repository, i.e. "rm
$BUP_DIR/refs/heads/some-branch", to remove branches/saves, etc. (see
test-gc.sh).

Thanks to Ben Kelly for help devising this Bloom-based approach, and for
testing and reporting bugs in earlier versions.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>