Add "bup rm" (UNSAFE)
Allow the removal of branches, and the removal of saves from specified
branches.  This command only removes the references, so until "bup rm"
is available, all of the related data will still be in the repository.

This command is potentially dangerous, so until we've had broader
testing, require all invocations to specify --unsafe, and make it clear
in the documentation that this command isn't considered stable.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>