Add git.object_info(objects), returning (hash, type, size) for each object.
This will allow improvements to the VFS, which currently crashes if
the repository contains a non-commit tag, and will support "bup get".

Even though we won't always want the size, use the default git
cat-file --batch-check format and always return the size, because
older versions of git don't support custom formats, and retrieving the
size shouldn't be notably more expensive (since it should be adjacent
to the type in the git object format).

Thanks to Alexander Barton <alex@barton.de> for reporting the git
incompatibility in an earlier version of this patch.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>