Add bup get; see the documentation for further information.
WARNING: this is preliminary, don't trust it yet.

Thanks to Karl Kiniger <karl.kiniger@med.ge.com> for helping track
down various bugs in earlier versions, and for noting that we might
want --verbose to be a little more effusive.  And thanks to Patryck
Rouleau <prouleau72@gmail.com> for suggesting improvements to the
documentation.

[rlb@defaultvalue.org: v8 remove some get-cmd.py code that's been
 adopted in master; add --git-dir to git-cat-tree; use git-cat-tree in
 test-get.sh; change walk_objects() to return hex ids, matching the
 input argument]

[git@tim-riemenschneider.de: rebased to current master and adjust to
 version of walk_object adopted in master]
Signed-off-by: Rob Browning <rlb@defaultvalue.org>